resonance-fixed1/backend/requirements.txt
fastapi
uvicorn[standard]
google-generativeai
firebase-admin
python-dotenv
cryptography
pydantic
python-multipart
